refactor all slackrtm structure into something sane
todo: split up slack command handler into own file "commands_slack"
